Convert SwapItemFormatter to SwapViewItemHelper
